ref: Move store body extractor to extractors
